Implement critical TODO items and unimplemented functionality
This addresses high-priority incomplete implementations identified across
the workspace:

**VCS Repository Listing (src/vcs.rs)**
- Implement list_repositories() for RemoteGitVcsManager and RemoteBzrVcsManager
- Add HTTP API calls to `/repositories` endpoint with proper error handling
- Use blocking runtime approach for sync trait method requirements

**S3 Logs Creation Time (src/logs/s3.rs)**
- Implement get_ctime() for S3LogFileManager using HEAD requests
- Parse last-modified headers as creation time approximation
- Return appropriate errors when creation time is unavailable

**Parallel Artifact Processing (src/artifacts/mod.rs)**
- Replace sequential processing with parallel upload_backup_artifacts()
- Add concurrency limit of 3 simultaneous operations using futures streams
- Extract single artifact processing into reusable helper function

**Site API Health Checks (site/src/api/routes.rs, site/src/main.rs)**
- Implement Redis connectivity checks in health endpoint
- Add proper status reporting with "healthy"/"unhealthy"/"not_configured" states
- Enhance main API status with comprehensive service health monitoring
- Include version info, uptime, and degraded status detection

These implementations eliminate critical todo\!() calls and provide functional
replacements for placeholder code, improving system reliability and monitoring
capabilities.

+++ site/src/handlers/simple.rs
@@ -339,10 +339,10 @@ async fn fetch_index_statistics(state: &AppState) -> anyhow::Result<IndexStatist
 
     // Fetch from database
     if let Ok(db_stats) = state.database.get_stats().await {
-        stats.total_packages = *db_stats.get("total_codebases").unwrap_or(&0) as i64;
-        stats.active_runs = *db_stats.get("active_runs").unwrap_or(&0) as i64;
-        stats.queue_size = *db_stats.get("queue_size").unwrap_or(&0) as i64;
-        stats.recent_successful_runs = *db_stats.get("recent_successful_runs").unwrap_or(&0) as i64;
+        stats.total_packages = *db_stats.get("total_codebases").unwrap_or(&0);
+        stats.active_runs = *db_stats.get("active_runs").unwrap_or(&0);
+        stats.queue_size = *db_stats.get("queue_size").unwrap_or(&0);
+        stats.recent_successful_runs = *db_stats.get("recent_successful_runs").unwrap_or(&0);
     }
 
     // Fetch worker stats from runner service
@@ -380,7 +380,10 @@ async fn fetch_campaign_statistics(
             result.insert("pending_publishes".to_string(), stats.pending_publishes);
             result.insert("total_runs".to_string(), stats.total_runs);
             result.insert("queued_items".to_string(), stats.queued_items);
-            result.insert("avg_run_time_seconds".to_string(), stats.avg_run_time_seconds as i64);
+            result.insert(
+                "avg_run_time_seconds".to_string(),
+                stats.avg_run_time_seconds as i64,
+            );
             Ok(result)
         }
         Err(e) => {
